docker-compose/second-project/flask-app/requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
mysql-connector-python==8.0.30
